+++ XBrainLab/ui/widget/__init__.py
@@ -1,0 +1,4 @@
+from .card import CardWidget as CardWidget
+from .placeholder import PlaceholderWidget as PlaceholderWidget
+from .plot_figure_window import PlotFigureWindow as PlotFigureWindow
+from .single_plot_window import SinglePlotWindow as SinglePlotWindow